% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{FrissC3ChartZoom}
\alias{FrissC3ChartZoom}
\title{Zooms an existing C3Chart from shiny}
\usage{
FrissC3ChartZoom(session, outputId, minX, maxX)
}
\arguments{
\item{outputId}{Id of the C3LineBarChart to update}

\item{minX}{minimum X value of the zoomed region}

\item{maxX}{maximum X value of the zoomed region}

\item{Shiny}{session id}
}
\description{
Zooms an existing C3Chart from shiny
}
