% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_area_chart.R
\name{p3_area_chart}
\alias{p3_area_chart}
\title{Area Chart}
\usage{
p3_area_chart(
  dataset,
  colors,
  chart_type = "area",
  axis_regions = NULL,
  axis_labels = NULL,
  labels_pos = NULL,
  show_points = FALSE,
  show_y2 = FALSE,
  zoom = TRUE,
  subchart = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors specifying the datasets to be plotted.}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{chart_type}{Character value for the chart type to plot e.g 'area' or line.
Defaults to area line chart}

\item{axis_regions}{Named list of character list defining the properties of regions to be shadded}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels.}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{show_points}{Boolean option to show data points on the chart plot}

\item{show_y2}{Boolean option to include a second y-axis.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\description{
Display as Area Chart.
}
\examples{

dataset <- list(data1=c(300, 350, 300, 0, 0, 0),
                data2=c(130, 100, 140, 200, 150, 50))
axis_regions <- list(list(axis='x',start= 1,end= 2,reg_class='regionX'),
                     list(axis='x',start= 4,end= 5,reg_class='regionX'))
colors <- list(data1="orange",data2="green")
axis_labels <- list(x_axis="species",y_axis="frequency")
labels_pos  <- list(xpos="outer-center",ypos="outer-middle")

 p3_area_chart(dataset,colors,chart_type='area',axis_regions,axis_labels=axis_labels,labels_pos=labels_pos)
 p3_area_chart(dataset,colors,chart_type='line',axis_regions,axis_labels,labels_pos)
}
