% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_combination_chart.R
\name{p3_combination_chart}
\alias{p3_combination_chart}
\title{Combinational Chart}
\usage{
p3_combination_chart(
  dataset,
  type,
  chart_types,
  data_groups = NULL,
  axis_regions = NULL,
  colors = NULL,
  show_points = FALSE,
  axis_labels = NULL,
  labels_pos = NULL,
  axis_rotate = FALSE,
  subchart = FALSE,
  zoom = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors containing the datasets to be plotted}

\item{type}{Character value defining the default chart type for all datasets. e.g type='bar'}

\item{chart_types}{List of  character vectors specifying chart types for specific datasets. For datasets not
specified in the chart_types , the default type will be
applied}

\item{data_groups}{List of  character vectors defining different groupings for the  stacked chart.}

\item{axis_regions}{Named list of character list defining the properties of regions to be shadded}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{show_points}{Boolean option to show data values on the plot}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Combinational Chart for a given dataset
}
\description{
Plots a combination of different charts on the same axis
}
\examples{
dataset <- list(
data1=c(30, 20, 50, 40, 60, 50),
data2=c(200, 130, 90, 240, 130, 220),
data3=c(300, 200, 160, 400, 250, 250),
data4=c(200, 130, 90, 240, 130, 220),
data5=c(130, 120, 150, 140, 160, 150),
data6=c(90, 70, 20, 50, 60, 120))

axis_labels <- list(x_axis="species",y_axis="frequency")
axis_regions <- list(list(axis='x',start= 0,end= 2,reg_class='regionX'),
                     list(axis='x',start= 4,end= 5,reg_class='regionX'))
colors <- list(data1="orange",data2="green",data3="red")
data_groups <- list(grp1=c('data1','data2','data5'))
chart_types <- list(data3='spline',data4='line')
labels_pos <- list(xpos="outer-center",ypos="outer-middle")


p3_combination_chart (dataset,type='bar',chart_types,data_groups,show_points=TRUE,labels_pos=labels_pos)
p3_combination_chart (dataset,type='bar',chart_types,data_groups,axis_regions,colors,axis_labels=axis_labels,labels_pos=labels_pos,subchart=TRUE)
}
