% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_cptrs_bar.R, R/p3_spcs_rai_bar.R
\name{p3_cptrs_bar-shiny}
\alias{p3_cptrs_bar-shiny}
\alias{p3_cptrs_barOutput}
\alias{renderp3_cptrs_bar}
\alias{p3_spcs_rai_barOutput}
\alias{renderp3_spcs_rai_bar}
\title{Shiny bindings for p3_cptrs_bar}
\usage{
p3_cptrs_barOutput(outputId, width = "100\%", height = "400px")

renderp3_cptrs_bar(expr, env = parent.frame(), quoted = FALSE)

p3_spcs_rai_barOutput(outputId, width = "100\%", height = "400px")

renderp3_spcs_rai_bar(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a p3_cptrs_bar}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using p3_cptrs_bar within Shiny
applications and interactive Rmd documents.

Output and render functions for using p3_cptrs_bar within Shiny
applications and interactive Rmd documents.
}
