% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfp3_dshbrd_charts.R, R/p3_dshbrd_charts.R
\name{p3_dshbrd_line_bar_chart}
\alias{p3_dshbrd_line_bar_chart}
\title{p3_dshbrd_line_bar_chart}
\usage{
p3_dshbrd_line_bar_chart(
  timeData,
  counts,
  lineData,
  color,
  showSubChart = FALSE,
  timeFormat = "\%Y-\%m-\%d",
  debounce = 250,
  width = "100\%",
  height = "400px",
  transition = 250,
  yLabel1 = "y-label",
  yLabel2 = "y-label2",
  xLabel = "x-label"
)

p3_dshbrd_line_bar_chart(
  timeData,
  counts,
  lineData,
  color,
  showSubChart = FALSE,
  timeFormat = "\%Y-\%m-\%d",
  debounce = 250,
  width = "100\%",
  height = "400px",
  transition = 250,
  yLabel1 = "y-label",
  yLabel2 = "y-label2",
  xLabel = "x-label"
)
}
\description{
Friss custom c3.js HTMLWidget output binding
}
