% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{p3_dshbrd_line_chart}
\alias{p3_dshbrd_line_chart}
\title{Friss C3 LineChart}
\usage{
p3_dshbrd_line_chart(
  timeData,
  lineData,
  color,
  showSubChart = FALSE,
  timeFormat = "\%Y-\%m-\%d",
  debounce = 250,
  width = "100\%",
  height = "400px",
  transition = 250,
  yLabel = "y-label",
  xLabel = "x-label"
)
}
\description{
Friss custom c3.js HTMLWidget output binding
}
