% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{p3_dshbrd_line_chartOutput}
\alias{p3_dshbrd_line_chartOutput}
\title{C3 chart output function for use in Shiny}
\usage{
p3_dshbrd_line_chartOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{outputId of the component}

\item{width}{width of the chart}

\item{height}{height of the chart}

\item{chart}{charttype: either: p3_dshbrd_guage_chart /}
}
\description{
C3 chart output function for use in Shiny
}
