% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{p3_dshbrd_pie_chart}
\alias{p3_dshbrd_pie_chart}
\title{Friss C3 pie or donut chart}
\usage{
p3_dshbrd_pie_chart(
  value = data.frame(group1 = 20, group2 = 50, group3 = 70),
  type = "pie",
  dataHidden = list(),
  width = NULL,
  height = 250,
  maxGroups = NA,
  maxStringSize = 10,
  legendPosition = "bottom",
  transition = 250,
  displayValues = FALSE
)
}
\arguments{
\item{value}{dataframe with values for the pie, the dataframe should consists of one row with a column for each group.}

\item{type}{type of the chart, either donut or pie}

\item{width}{width of the div the gauge is put in}

\item{height}{height of the div the gauge is put in}
}
\description{
Friss C3 pie or donut chart
}
\examples{
\dontrun{
data <- data.frame(group1=20,group2=50,group3=70)
p3_dshbrd_pie_chart(data)
}
}
