% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_heatmap_chart.R
\name{p3_heatmap_chart}
\alias{p3_heatmap_chart}
\title{heatmap chart}
\usage{
p3_heatmap_chart(dataset, data_names = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{dataset}{List of vectors containing the datasets to be
plotted.}

\item{data_names}{List of characters defining the actual names of each data
defined in the dataset}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Heatmap Chart Plot
}
\description{
Plots a heatmap Chart for a specified set of data with color mapped value.
}
\examples{

data_names <- c(groups="Species",variables="Station",values="Images")
data <- data.frame(
  groups = sort(
    rep(
      c("Aardwolf","Serval","Hyaena_Brown","Civet_African","Leopard_African",
        "Jackal_Black_Backed","Lion_African","Cheetah","Genet_Large_Spotted",
        "Hyaena_Spotted"
      ),
      10
    )
  ),
  variables = rep(
    c("StationStation1","StationStation2","StationStation3","StationStation4",
      "StationStation5","StationStation6","StationStation7","StationStation8",
      "StationStation9","StationStation10"
    ),
    10
  ),
  values = round(
    runif(100,min=0,max=10000),
    0
  )
)
p3_heatmap_chart(dataset=data,data_names=data_names)
p3_heatmap_chart(data)

}
