% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_multiple_xy_chart.R
\name{p3_multiple_xy_chart}
\alias{p3_multiple_xy_chart}
\title{Multiple XY Line Chart}
\usage{
p3_multiple_xy_chart(
  dataset,
  colors = NULL,
  show_points = FALSE,
  axis_regions = NULL,
  axis_labels = NULL,
  labels_pos = NULL,
  axis_rotate = FALSE,
  subchart = FALSE,
  zoom = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors specifying the x-axis values and the corresponding
data points for the specified xaxis values.Order is very important, see example for details.}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{show_points}{Boolean option to show data points on the chart plot}

\item{axis_regions}{Named list of character lists defining the properties of regions to be shadded}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels.}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Spline Chart with specified parms.
}
\description{
Plots multiple line chart with multiple custom x values
}
\examples{
dataset <- list(x1=c(10, 30, 45, 50, 70, 100),
                x2=c(30, 50, 75, 100, 120),
                data1=c(30, 200, 100, 400, 150, 250),
                data2=c(20, 180, 240, 100, 190))

colors <- list(data1="orange",data2="green",data3="red")
axis_labels <- list(x_axis="species",y_axis="frequency")
labels_pos <- list(xpos="outer-center",ypos="outer-middle")
axis_regions <- list(list(axis='x',start= 30,end= 50,reg_class='regionX'),
                     list(axis='x',start= 75,end= 100,reg_class='regionX'))



p3_multiple_xy_chart(dataset,colors,show_points=TRUE)
p3_multiple_xy_chart (dataset,colors,FALSE,axis_regions,axis_labels,labels_pos,subchart=TRUE)
p3_multiple_xy_chart (dataset,colors,FALSE,axis_labels=axis_labels,labels_pos=labels_pos,subchart=TRUE)
}
