% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_scatter_chart.R
\name{p3_scatter_chart}
\alias{p3_scatter_chart}
\title{Scatter Chart}
\usage{
p3_scatter_chart(
  dataset,
  colors = NULL,
  axis_labels = NULL,
  labels_pos = NULL,
  axis_regions = NULL,
  subchart = FALSE,
  zoom = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors containing the datasets to be plotted}

\item{colors}{A list of plot colors preffred for each dataset}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_regions}{Named list of character lists defining the properties of regions to be shadded}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and}

\item{width, height}{Fixed width for widget (in css units). The default is
NULL, which results in intelligent automatic sizing
based on  the widget’s container.}

\item{elementId}{Use an explicit element ID for the widget. Useful if you have
other JavaScript that needs to explicitly discover and
interact with a specific widget instance. Default NULL
which results in an automatically generated one.}
}
\description{
Plots a Scatter Chart using Cartesian coordinates
                      to display values of a given dataset
}
\examples{
dataset <- list(
data1=c(30, 20, 50, 40, 60, 50),
data2=c(200, 130, 90, 240, 130, 220),
data3=c(300, 200, 160, 400, 250, 250))
axis_labels <- list(x_axis="Petals",y_axis="Sepals")
labels_pos <- list(xpos="outer-center",ypos="outer-middle")
colors <- list(data1="blue",data2="black",data3="red")
axis_regions <- list(list(axis='x',start= 1, end= 2,reg_class='regionX'),
                     list(axis='x',start=4, end=5,reg_class='regionX'))



p3_scatter_chart(dataset,colors,axis_labels,labels_pos,axis_regions,TRUE,TRUE)
p3_scatter_chart(dataset,NULL,NULL,NULL,axis_regions,FALSE,TRUE,'80\%','200\%')
p3_scatter_chart(dataset,colors,axis_labels,labels_pos)
p3_scatter_chart(dataset,colors,axis_labels)
\dontrun{

 p3_scatter_chart(dataset,colors,'100\%','500\%')
}

}
