% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_spcs_rai_bar.R
\name{p3_spcs_rai_bar}
\alias{p3_spcs_rai_bar}
\title{p3_cptrs_bar}
\usage{
p3_spcs_rai_bar(
  dataset,
  colors,
  show_values = TRUE,
  plot_type = "bar",
  axis_rotate = TRUE,
  show_y2 = FALSE,
  zoom = TRUE,
  subchart = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors specifying the x-axis values and the
corresponding data points for the specified x-axis values.
Order is very important, see example for details.}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{show_values}{Boolean option to show data values on all bars.
Defaults to TRUE.}

\item{plot_type}{Character value specifying the chart type for the plot
e.g bar , line etc. Defaults to bar chart.}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{show_y2}{Boolean value to include a second y-axis. Default is set
to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\description{
Bar chart plot for the number of camera trap images captured per specified species
}
\examples{
dataset <- list(Species=c("Lion","Baboon","Tiger","Monkey","Leopard","Bird",
                         "Girrafe","Kudu","Springbook","Snake"),
                         Captures=c(20, 180, 240, 100, 190,213,45,189,67,1))
colors <- list(data1="orange",data2="green",data3="red")
p3_spcs_rai_bar(dataset = dataset , colors = colors)
p3_spcs_rai_bar(dataset,colors,plot_type='line',axis_rotate=FALSE)
p3_spcs_rai_bar(dataset,colors,axis_rotate=TRUE)

}
