% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_area.R
\name{p3_stacked_area}
\alias{p3_stacked_area}
\title{Stacked Area Chart}
\usage{
p3_stacked_area(dataset, colors = NULL, show_points = FALSE,
  axis_labels = NULL, labels_pos = NULL, subchart = TRUE,
  zoom = TRUE, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{dataset}{List of numeric vectors containing the datasets to be plotted
and a factor vector specifying the time category}

\item{colors}{A list of plot colors prefered for each dataset}

\item{show_points}{Boolean option to show data values on the chart plot}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels.}

\item{labels_pos}{Named list of characters defining the prefered position
of the axis labelse.g for x-axis ( inner-center, inner-left,
outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top,
outer-middle, outer-bottom, inner-top [default] )}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\description{
Extension of a basic area chart to display the evolution of the value of
several groups on the same graphic.
}
\examples{
dataset <- list(
Time=c("2016-01-05","2016-01-12","2016-01-24","2016-02-05","2016-02-12","2016-02-24"),
data1=c(200, 130, 90, 240, 130, 220),
data2=c(300, 200, 160, 400, 250, 250),
data3=c(10, 20, 50, 40, 80, 65))

colors <- list(data1="orange",data2="green",data3="red")
axis_labels <- list(x_axis="species",y_axis="frequency")
labels_pos <- list(xpos="outer-center",ypos="outer-middle")

p3_stacked_area(dataset, colors)
p3_stacked_area(dataset, colors,axis_labels=axis_labels,labels_pos=labels_pos)
p3_stacked_area(dataset, colors, TRUE,axis_labels,labels_pos,TRUE,TRUE)
}
