% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_bar_chart.R
\name{p3_stacked_bar_chart}
\alias{p3_stacked_bar_chart}
\title{Stacked Bar Chart}
\usage{
p3_stacked_bar_chart(dataset, data_groups, x_categories = NULL,
  colors = NULL, axis_labels = NULL, labels_pos = NULL,
  axis_rotate = FALSE, subchart = FALSE, zoom = TRUE, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{dataset}{List of numeric vectors specifying the datasets to be plotted.}

\item{data_groups}{List of character vectors defining different groupings for the chart.}

\item{x_categories}{Character vector of names for the x category (Show ticks as categorized by each data.)}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Stacked Bar Chart with specified parms.
}
\description{
Plots a series of bar charts stacked according to specified data_groups
}
\examples{
dataset <- list(
data1=c(30, 20, 50, 40, 60, 50),
data2=c(200, 130, 90, 240, 130, 220),
data3=c(300, 200, 160, 400, 250, 250),
data4=c(200, 130, 90, 240, 130, 220))

data_groups <- list(grp1=c('data1','data3'),grp2=c('data2','data4'))
x_categories <- c('one','two','three','four','five','six')
colors <- list(data1="orange",data2="green",data3="red")
axis_labels <- list(x_axis="species",y_axis="frequency")
labels_pos <- list(xpos="outer-center",ypos="outer-middle")

p3_stacked_bar_chart(dataset,data_groups,x_categories,colors,labels_pos=labels_pos)
p3_stacked_bar_chart (dataset,data_groups,x_categories,colors,axis_labels,labels_pos,subchart=TRUE)

}
