% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{p3_stacked_chart}
\alias{p3_stacked_chart}
\title{p3_stacked_chart}
\usage{
p3_stacked_chart(dataset, colors, types = NULL, groups = NULL,
  subchart = FALSE, axis_type = "timeseries", categories = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{dataset}{todo}

\item{colors}{named list given series colors e.g. list(data1 = "purple", data2 = "blue", data3 = "gray")}

\item{types}{named list given series chart types e.g. list(data1 = "bar", data2 = "bar", data3 = "bar")}

\item{groups}{vector given stacking order e.g. c("data1","data2","data3")}

\item{subchart}{boolean, should a subchart be show with a brushable area?}

\item{axis_type}{what type of x-axis should be drawn, either timeseries, category or indexed, defaults to timeseries}

\item{categories}{category labels}

\item{width}{chart width}

\item{height}{chart height}
}
\description{
p3_stacked_chart
}
