% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_step_chart.R
\name{p3_step_chart}
\alias{p3_step_chart}
\title{Step Chart}
\usage{
p3_step_chart(
  dataset,
  chart_types = NULL,
  show_points = FALSE,
  colors = NULL,
  axis_labels = NULL,
  labels_pos = NULL,
  axis_rotate = FALSE,
  subchart = FALSE,
  zoom = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors specifying the datasets to be
plotted}

\item{chart_types}{List of character vectors specifying chart types for
specific datasets .e.g area-step,line .For datasets not
specified in the chart_types , the default (line) type will be
applied}

\item{show_points}{Boolean option to show data points on the chart plot.}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels.}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Combinational Chart for a given dataset
}
\description{
Plots a line based step chart usefull in showing data that doesn't changes
very often.
}
\examples{
dataset <- list(
data1=c(30, 20, 50, 40, 60, 50),
data2=c(200, 190, 220, 210, 400, 220)
)

axis_labels <- list(x_axis="species",y_axis="frequency")
colors <- list(data1="orange",data2="green")
chart_types <- list(data1='area-step',data2='line')
labels_pos <- list(xpos="outer-center",ypos="outer-middle")


p3_step_chart (dataset,chart_types)
p3_step_chart (dataset,chart_types,colors=colors,axis_labels=axis_labels,labels_pos=labels_pos,subchart=TRUE)
}
