% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_time_series_chart.R
\name{p3_time_series_chart}
\alias{p3_time_series_chart}
\title{TimeSeries Chart}
\usage{
p3_time_series_chart(
  dataset,
  colors = NULL,
  axis_regions = NULL,
  show_points = FALSE,
  time_format = NULL,
  axis_labels = NULL,
  labels_pos = NULL,
  subchart = FALSE,
  zoom = TRUE,
  show_y2 = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors containing the datasets to be plotted and a factor vector specifying the time category}

\item{colors}{A list of plot colors prefered for each dataset}

\item{axis_regions}{Named list of character lists defining the properties of regions to be shadded}

\item{show_points}{Boolean option to show data values points}

\item{time_format}{Character value specifying the display time format .}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels.}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels.
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
 and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{subchart}{Boolean option to show sub chart for zoom and selection range.Default set to False..}

\item{zoom}{Boolean option to Zoom by mouse wheel event and slide by drag. Default set to True}

\item{show_y2}{Boolean option to show a second y-axis on the chart axis.}

\item{width, height}{Fixed width for widget (in css units).The default is NULL
which results in intelligent automatic sizing based on the
widget’s container.}

\item{elementId}{Use an explicit element ID for the widget. Useful if you have
other JavaScript that needs to explicitly discover and
interact with a specific widget instance. Default NULL
which results in an automatically generated one.}
}
\value{
TimeSeries Chart
}
\description{
Plots data trends in numeric data over a time period.
}
\examples{
dataset <- data.frame(
    Time=c('2013-01-01','2013-01-02','2013-01-03','2013-01-04','2013-01-05','2013-01-06'),
    data1=c(30, 200, 100, 400, 150, 250),
    data2=c(130, 340, 200, 500, 250, 350),
    data3=c(400, 500, 450, 700, 600, 500))
colors <- list(data1="orange",data2="green",data3="red")
axis_regions <- list(list(start= '2013-01-02', end= '2013-01-03'),
                     list(start='2013-01-05', end='2013-01-06'))
labels_pos <- list(xpos="outer-center",ypos="outer-middle")

p3_time_series_chart(dataset,colors,axis_regions,labels_pos=labels_pos)
p3_time_series_chart(dataset,colors,axis_regions,show_points=TRUE,labels_pos=labels_pos,subchart=TRUE,zoom=TRUE)
}
