% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_xy_line_chart.R
\name{p3_xy_line_chart}
\alias{p3_xy_line_chart}
\title{Simple line chart}
\usage{
p3_xy_line_chart(
  dataset,
  colors = NULL,
  axis_regions = NULL,
  show_points = FALSE,
  axis_labels = NULL,
  labels_pos = NULL,
  axis_rotate = FALSE,
  subchart = FALSE,
  zoom = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dataset}{List of numeric vectors containing the datasets to be plotted.
'x' named vector specifies the x-axis values for the plot.}

\item{colors}{Named list with colors for the data series in the chart.
NULL results in an random automatically generated colors.}

\item{axis_regions}{Named list of character lists defining the properties of regions to be shadded.}

\item{show_points}{Boolean option to show data points on the chart plot}

\item{axis_labels}{Named list of characters defining the prefered chart axis labels}

\item{labels_pos}{Named list of characters defining the prefered position of the axis labels
e.g for x-axis ( inner-center, inner-left, outer-right, outer-center, outer-left, inner-right [default] )
and  y-axis ( inner-middle, inner-bottom, outer-top, outer-middle, outer-bottom, inner-top [default] )}

\item{axis_rotate}{Boolean value to determine axis rotation. Default is set
to False.}

\item{subchart}{Boolean option to show sub chart for zoom and selection
range.Default set to False.}

\item{zoom}{Boolean option to Zoom by mouse wheel event and
slide by drag. Default set to True}

\item{width, height}{Must be a valid CSS unit (like '100%', '400px', 'auto')
or a number, which will be coerced to a string and have
'px' appended.The default is NULL, which results in
intelligent automatic sizing based on the chart’s
container.}

\item{elementId}{Use an explicit element ID for the widget Useful if you
have other JavaScript that needs to explicitly discover
and interact with a specific widget instance. in any
other case leave as NULL which results in an
automatically generated one.}
}
\value{
Line chart with custom x.
}
\description{
Simple line chart with custom x vlaues
}
\examples{
dataset <- list(
x=c(30, 50, 100, 230, 300, 310),
data1=c(30, 200, 100, 400, 150, 250),
data2=c(130, 300, 200, 300, 250, 450))

colors <- list(data1="green",data2="red")
axis_regions <- list(list(axis='x',start= 60,end= 90,reg_class='regionX'),
                     list(axis='x',start= 181,end= 222,reg_class='regionX'))
axis_labels <- list(x_axis="species",y_axis="frequency")
labels_pos <- list(xpos="outer-center",ypos="outer-middle")

p3_xy_line_chart(dataset,colors,axis_regions,axis_labels=axis_labels,labels_pos=labels_pos,subchart=TRUE)
}
