% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{renderP3_dshbrd_charts}
\alias{renderP3_dshbrd_charts}
\title{Widget render function for use in Shiny}
\usage{
renderP3_dshbrd_charts(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
