% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_dshbrd_charts.R
\name{renderP3_dshbrd_guage_chart}
\alias{renderP3_dshbrd_guage_chart}
\title{C3 gauge Widget render function for use in Shiny}
\usage{
renderP3_dshbrd_guage_chart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
C3 gauge Widget render function for use in Shiny
}
