% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{setColors}
\alias{setColors}
\title{Sets the colors of an existing PantheraWidgets chart}
\usage{
setColors(chartProxy, colors)
}
\arguments{
\item{chartProxy}{PantheraWidgets chart proxy object}

\item{colors}{named list with colors for the data series in the chart}
}
\value{
PantheraWidgets proxy object
}
\description{
<Add Title>
}
\details{
<Add Description>
}
