% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{setGroups}
\alias{setGroups}
\title{Set chart groups}
\usage{
setGroups(chartProxy, groups)
}
\arguments{
\item{chartProxy}{PantheraWidgets chart proxy object}

\item{groups}{vector containing groups used to stack the chart}
}
\value{
PantheraWidgets proxy object
}
\description{
Set chart groups
}
