% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{setTypes}
\alias{setTypes}
\title{Set chart types}
\usage{
setTypes(chartProxy, info)
}
\arguments{
\item{chartProxy}{PantheraWidgets chart proxy object}

\item{message}{list containing data series types}
}
\value{
PantheraWidgets proxy object
}
\description{
<Add Title>
}
\details{
<Add Description>
}
