% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sngl_img_viewer.R
\name{sngl_img_viewer-shiny}
\alias{sngl_img_viewer-shiny}
\alias{sngl_img_viewerOutput}
\alias{renderSngl_img_viewer}
\title{Shiny bindings for sngl_img_viewer}
\usage{
sngl_img_viewerOutput(outputId, width = "100\%", height = "400px")

renderSngl_img_viewer(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a sngl_img_viewer}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using sngl_img_viewer within Shiny
applications and interactive Rmd documents.
}
