% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sngl_img_viewer.R
\name{sngl_img_viewer}
\alias{sngl_img_viewer}
\title{Single Image Viewer}
\usage{
sngl_img_viewer(filename, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{width, height}{Should be set to 0. The targeted element width and height
will be used instead.}

\item{elementId}{Use an explicit element ID where the viewer will be
displayed.This can be the ID of a div element on your
page.}
}
\description{
Renders a single image viewer
}
