% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{transformTo}
\alias{transformTo}
\title{Set all chart types simulaneously}
\usage{
transformTo(chartProxy, type)
}
\arguments{
\item{chartProxy}{PantheraWidgets chart proxy object}

\item{message}{list containing data series type}
}
\value{
PantheraWidgets proxy object
}
\description{
Set all chart types simulaneously
}
