% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_stacked_chart.R
\name{zoomChart}
\alias{zoomChart}
\title{Zooms an existing PantheraWidgetsChart from shiny}
\usage{
zoomChart(chartProxy, minX, maxX)
}
\arguments{
\item{minX}{date start}

\item{maxX}{date end}
}
\value{
PantheraWidgets proxy object
}
\description{
<Add Title>
}
\details{
<Add Description>
}
